refactor(componentLoading): move view factory access to the router view port
Previously, the component instantiation code was placed inside the
component loading step. This update moves that code into the
RouterViewPort. This decreases the number of dependencies that component
loading has, but it also brings an additional benefit to the routing
system. By allowing the RouterViewPort to control component
instantiation, it can optionally choose to cache components internally
in a way that could improve performance for frequently visiting pages.
This caching is not implemented as part of this update, but it's now
possible to add it in the future.