refactor(pipelineProvider): enable pipeline steps to be created through di
Previously, the PipelineProvider was manually instantiating each step in
the navigation pipeline. After a recent refactor this became a problem
since the LoadNewComponentsStep had a dependency on templating's
ComponentLoader. This was temporarily solved by having the provider push
that into the step manually, but that created a coupling between the
provider and its individual step's dependencies. As of this fix, the
provider instances each step via the injector, allowing any pipeline
step to declare it own dependencies without the provider needing to know
about them.